feat(redirects and pipes): :sparkles: fix redirects and pipes and norminette
add the redirect and pipes and fixed norminette, also changed the main scope, now we can do the execution
