+++ minishell/src/utilits_main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utilits_main.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gda-conc <gda-conc@student.42.fr>          +#+  +:+       +#+        */
+/*   By: bieldojt <bieldojt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/15 15:46:09 by guclemen          #+#    #+#             */
-/*   Updated: 2025/05/19 14:44:22 by gda-conc         ###   ########.fr       */
+/*   Updated: 2025/05/21 16:14:20 by bieldojt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,4 +27,37 @@ void	ft_clean_shell(t_shell *shell)
 	shell->input = NULL;
 	shell->tokens = NULL;
 	shell->cmds = NULL;
+}
+
+int	is_space_or_invalid(char *input)
+{
+	if (!ft_not_only_spaces(input))
+	{
+		free(input);
+		input = NULL;
+		return (1);
+	}
+	if (check_syntax_error(input))
+	{
+		free(input);
+		input = NULL;
+		return (1);
+	}
+	return (0);
+}
+
+t_token	*list_token(char *input)
+{
+	t_token	*token_list;
+
+	token_list = lexer(input);
+	expand_variables_in_token(token_list);
+	clean_tokens(token_list);
+	return (token_list);
+}
+
+void	free_input_token_cmd(char *input, t_cmd *cmd_list)
+{
+	free(input);
+	free_commands(cmd_list);
 }
